Separate GenericObjectSerializer and GenericObjectValidator
