+++ Assets/VRM/UniJSON/Scripts/JsonSchemaValidator/GenericObjectSerializer.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 3618a154b60ca454ebcc0bfeeb038fbe
+timeCreated: 1550086962
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
