+++ Assets/VRM/UniJSON/Scripts/JsonSchemaValidator/GenericObjectValidator.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 3cdcb3f68ea0623469e7943cf5f2b843
+timeCreated: 1550086731
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
