+++ Assets/VRM/UniJSON/Scripts/JsonSchemaValidator/JsonObjectValidator.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Reflection;
 
 
 namespace UniJSON
@@ -297,178 +296,6 @@ namespace UniJSON
             }
         }
 
-        static class GenericFieldView<T>
-        {
-            public static FieldInfo[] GetFields()
-            {
-                var t = typeof(T);
-                return t.GetFields(BindingFlags.Instance | BindingFlags.Public);
-            }
-
-            public static void CreateFieldProcessors<G, D>(
-                Func<FieldInfo, D> creator,
-                Dictionary<string, D> processors
-                )
-            {
-                foreach (var fi in GetFields())
-                {
-                    processors.Add(fi.Name, creator(fi));
-                }
-            }
-        }
-
-        internal class ValidationResult
-        {
-            public bool IsIgnorable;
-            public JsonSchemaValidationException Ex;
-        }
-
-        public static class GenericValidator<T>
-        {
-            class ObjectValidator
-            {
-                delegate JsonSchemaValidationException FieldValidator(
-                    JsonSchema s, JsonSchemaValidationContext c, T o, out bool isIgnorable);
-
-                Dictionary<string, FieldValidator> m_validators;
-
-                static FieldValidator CreateFieldValidator(FieldInfo fi)
-                {
-                    var mi = typeof(ObjectValidator).GetMethod("_CreateFieldValidator",
-                        BindingFlags.Static | BindingFlags.NonPublic)
-                        ;
-                    var g = mi.MakeGenericMethod(fi.FieldType);
-                    return GenericInvokeCallFactory.StaticFunc<FieldInfo, FieldValidator>(g)(fi);
-                }
-
-                static FieldValidator _CreateFieldValidator<U>(FieldInfo fi)
-                {
-                    var getter = (Func<T, U>)((t) => (U)fi.GetValue(t));
-
-                    return (JsonSchema s, JsonSchemaValidationContext c, T o, out bool isIgnorable) =>
-                    {
-                        var v = s.Validator;
-                        using (c.Push(fi.Name))
-                        {
-                            var field = getter(o);
-                            var ex = v.Validate(c, field);
-
-                            isIgnorable = ex != null && s.IsExplicitlyIgnorableValue(field);
-
-                            return ex;
-                        }
-                    };
-                }
-
-                public ObjectValidator()
-                {
-                    var validators = new Dictionary<string, FieldValidator>();
-                    GenericFieldView<T>.CreateFieldProcessors<ObjectValidator, FieldValidator>(
-                        CreateFieldValidator, validators);
-
-                    m_validators = validators;
-                }
-
-                public JsonSchemaValidationException ValidateProperty(
-                    HashSet<string> required,
-                    KeyValuePair<string, JsonSchema> property,
-                    JsonSchemaValidationContext c,
-                    T o,
-                    out bool isIgnorable
-                    )
-                {
-                    var fieldName = property.Key;
-                    var schema = property.Value;
-
-                    isIgnorable = false;
-
-                    FieldValidator fv;
-                    if (m_validators.TryGetValue(fieldName, out fv))
-                    {
-                        var isRequired = required != null && required.Contains(fieldName);
-
-                        bool isMemberIgnorable;
-                        var ex = fv(schema, c, o, out isMemberIgnorable);
-                        if (ex != null)
-                        {
-                            isIgnorable = !isRequired && isMemberIgnorable;
-
-                            if (isRequired // required fields must be checked
-                                || c.EnableDiagnosisForNotRequiredFields)
-                            {
-                                return ex;
-                            }
-                        }
-                    }
-
-                    return null;
-                }
-
-                public JsonSchemaValidationException Validate(
-                    HashSet<string> required,
-                    Dictionary<string, JsonSchema> properties,
-                    JsonSchemaValidationContext c, T o)
-                {
-                    foreach (var kv in properties)
-                    {
-                        bool isIgnorable;
-                        var ex = ValidateProperty(required, kv, c, o, out isIgnorable);
-                        if (ex != null && !isIgnorable)
-                        {
-                            return ex;
-                        }
-                    }
-
-                    return null;
-                }
-
-                public void ValidationResults
-                    (HashSet<string> required,
-                    Dictionary<string, JsonSchema> properties,
-                    JsonSchemaValidationContext c, T o,
-                    Dictionary<string, ValidationResult> results)
-                {
-                    foreach (var kv in properties)
-                    {
-                        bool isIgnorable;
-                        var ex = ValidateProperty(required, kv, c, o, out isIgnorable);
-
-                        results.Add(kv.Key, new ValidationResult {
-                            IsIgnorable = isIgnorable,
-                            Ex = ex,
-                        });
-                    }
-                }
-            }
-
-            static ObjectValidator s_validator;
-
-            static void prepareValidator()
-            {
-                if (s_validator == null)
-                {
-                    s_validator = new ObjectValidator();
-                }
-            }
-
-            public static JsonSchemaValidationException Validate(HashSet<string> required,
-                Dictionary<string, JsonSchema> properties,
-                JsonSchemaValidationContext c, T o)
-            {
-                prepareValidator();
-                return s_validator.Validate(required, properties, c, o);
-            }
-
-            internal static void ValidationResults(HashSet<string> required,
-                Dictionary<string, JsonSchema> properties,
-                JsonSchemaValidationContext c, T o,
-                Dictionary<string, ValidationResult> results)
-            {
-                prepareValidator();
-                s_validator.ValidationResults(required, properties, c, o, results);
-            }
-        }
-
         public JsonSchemaValidationException Validate<T>(JsonSchemaValidationContext c, T o)
         {
             if (o == null)
@@ -481,110 +308,7 @@ namespace UniJSON
                 return new JsonSchemaValidationException(c, "no properties");
             }
 
-            return GenericValidator<T>.Validate(Required, Properties, c, o);
-        }
-
-        static class GenericObjectSerializer<T>
-        {
-            class Serializer
-            {
-                delegate void FieldSerializer(JsonSchema s, JsonSchemaValidationContext c, IFormatter f, T o,
-                    Dictionary<string, ValidationResult> vRes, string[] deps);
-
-                Dictionary<string, FieldSerializer> m_serializers;
-
-                static FieldSerializer CreateFieldSerializer(FieldInfo fi)
-                {
-                    var mi = typeof(Serializer).GetMethod("_CreateFieldSerializer",
-                        BindingFlags.Static | BindingFlags.NonPublic);
-                    var g = mi.MakeGenericMethod(fi.FieldType);
-                    return GenericInvokeCallFactory.StaticFunc<FieldInfo, FieldSerializer>(g)(fi);
-                }
-
-                static FieldSerializer _CreateFieldSerializer<U>(FieldInfo fi)
-                {
-                    Func<T, U> getter = t =>
-                    {
-                        return (U)fi.GetValue(t);
-                    };
-
-                    return (s, c, f, o, vRes, deps) =>
-                    {
-                        var v = s.Validator;
-                        var field = getter(o);
-
-                        if (vRes[fi.Name].Ex != null)
-                        {
-                            return;
-                        }
-
-                        if (deps != null)
-                        {
-                            foreach(var dep in deps)
-                            {
-                                if (vRes[dep].Ex != null)
-                                {
-                                    return;
-                                }
-                            }
-                        }
-
-                        f.Key(fi.Name);
-                        v.Serialize(f, c, field);
-                    };
-                }
-
-                public Serializer()
-                {
-                    var serializers = new Dictionary<string, FieldSerializer>();
-                    GenericFieldView<T>.CreateFieldProcessors<Serializer, FieldSerializer>(
-                        CreateFieldSerializer, serializers);
-
-                    m_serializers = serializers;
-                }
-
-                public void Serialize(JsonObjectValidator objectValidator,
-                    IFormatter f, JsonSchemaValidationContext c, T o)
-                {
-                    // Validates fields
-                    var validationResults = new Dictionary<string, ValidationResult>();
-                    GenericValidator<T>.ValidationResults(
-                        objectValidator.Required, objectValidator.Properties,
-                        c, o, validationResults);
-
-                    // Serialize fields
-                    f.BeginMap(objectValidator.Properties.Count());
-                    foreach (var property in objectValidator.Properties)
-                    {
-                        var fieldName = property.Key;
-                        var schema = property.Value;
-
-                        string[] deps = null;
-                        objectValidator.Dependencies.TryGetValue(fieldName, out deps);
-
-                        FieldSerializer fs;
-                        if (m_serializers.TryGetValue(fieldName, out fs))
-                        {
-                            fs(schema, c, f, o, validationResults, deps);
-                        }
-                    }
-                    f.EndMap();
-                }
-            }
-
-            static FieldInfo[] s_fields;
-            static Serializer s_serializer;
-
-            public static void Serialize(JsonObjectValidator objectValidator,
-                    IFormatter f, JsonSchemaValidationContext c, T value)
-            {
-                if (s_serializer == null)
-                {
-                    s_serializer = new Serializer();
-                }
-
-                s_serializer.Serialize(objectValidator, f, c, value);
-            }
+            return GenericObjectValidator<T>.Validate(Required, Properties, c, o);
         }
 
         public void Serialize<T>(IFormatter f, JsonSchemaValidationContext c, T value)
